origin/12-f-redux-article-app: Rename folders
